refactor(mobile): split host client hooks
