+++ mobile/src/transport/client-context.tsx
@@ -7,10 +7,8 @@ import {
   useEffect,
   useMemo,
   useRef,
-  useState,
   type ReactNode
 } from 'react'
-import type { RpcClient } from './rpc-client'
 import type { StableLogicalRpcClient } from './stable-logical-rpc-client'
 import { subscribeConnectionRevivalTriggers } from './connection-revival-triggers'
 import { HostClientOpenRegistry } from './host-client-open-registry'
@@ -34,6 +32,15 @@ import {
 } from './host-client-context-state'
 import type { ConnectionState, HostProfile } from './types'
 import type { RpcClientContextValue } from './rpc-client-context-contract'
+
+export {
+  useDisconnectHostClient,
+  useForceReconnect,
+  useForgetHostClient,
+  useHostClient,
+  usePrimeHosts,
+  useRefreshHostClient
+} from './host-client-hooks'
 
 type StoreEntry = HostClientStoreEntry
 
@@ -363,107 +370,4 @@ export function useRpcClientContext(): RpcClientContextValue {
     throw new Error('useHostClient must be used inside <RpcClientProvider>')
   }
   return ctx
-}
-
-// Primary hook for screens: acquires the shared client on mount, releases on unmount, re-renders on state change.
-export function useHostClient(hostId: string | undefined): {
-  client: RpcClient | null
-  clientId: string | null
-  state: ConnectionState
-} {
-  const ctx = useRpcClientContext()
-  const [, force] = useState(0)
-  // Why: an absent entry at mount is almost always the open racing the render, not a
-  // dead host — seed amber; a failed open notifies 'disconnected' moments later.
-  const [state, setState] = useState<ConnectionState>(() =>
-    hostId ? (ctx.getKnownState(hostId) ?? 'connecting') : 'disconnected'
-  )
-  const clientRef = useRef<RpcClient | null>(null)
-  const clientHostIdRef = useRef<string | undefined>(hostId)
-  const acquisitionRef = useRef<HostClientAcquisition>({})
-
-  useEffect(() => {
-    if (!hostId) {
-      clientRef.current = null
-      clientHostIdRef.current = undefined
-      setState('disconnected')
-      return
-    }
-    clientHostIdRef.current = hostId
-    let cancelled = false
-    // Subscribe before acquire so any state change during open is captured.
-    const unsub = ctx.subscribeHostState(hostId, (next) => {
-      if (cancelled) {
-        return
-      }
-      setState(next)
-      // Why: async open and forceReconnect swap the client object; re-read each state change so screens never drive a stale one.
-      const found = ctx.getAllClients().find((entry) => entry.hostId === hostId)
-      if (found && found.client !== clientRef.current) {
-        clientRef.current = found.client
-        force((n) => n + 1)
-      } else if (!found && clientRef.current) {
-        // Why: disconnect/forget deletes the entry; never retain a dead client (STA-1511).
-        clientRef.current = null
-        force((n) => n + 1)
-      }
-    })
-    const initial = ctx.acquire(hostId, acquisitionRef.current)
-    clientRef.current = initial
-    setState(ctx.getKnownState(hostId) ?? 'connecting')
-    if (initial) {
-      // Why: two cached hosts can both be connected, so equal state values cannot reveal the replacement client.
-      force((n) => n + 1)
-    }
-    return () => {
-      cancelled = true
-      unsub()
-      ctx.release(hostId, acquisitionRef.current)
-      clientRef.current = null
-      clientHostIdRef.current = undefined
-    }
-  }, [ctx, hostId])
-
-  // Why: Expo can reuse the screen before effects bind the next host; never expose the prior host's client or state in that render.
-  const bound = clientHostIdRef.current === hostId
-  const boundClient = bound ? clientRef.current : null
-  const boundState = bound
-    ? state
-    : hostId
-      ? (ctx.getKnownState(hostId) ?? 'connecting')
-      : 'disconnected'
-  // Why: publish identity from the same entry as the client so consumers cannot race a second Keychain read.
-  return {
-    client: boundClient,
-    clientId: boundClient && hostId ? ctx.getClientId(hostId) : null,
-    state: boundState
-  }
-}
-
-// Why: host-store's removeHost() must close the live client but has no React-side handle; this hook bridges to it.
-export function useRefreshHostClient(): (hostId: string) => void {
-  const ctx = useRpcClientContext()
-  return ctx.refreshHostClient
-}
-
-export function useForgetHostClient(): (hostId: string) => void {
-  const ctx = useRpcClientContext()
-  return ctx.forgetHostClient
-}
-
-export function useDisconnectHostClient(): (hostId: string) => void {
-  const ctx = useRpcClientContext()
-  return ctx.disconnectHostClient
-}
-
-// Why: future-proof "Connection issues — try again" affordance.
-export function useForceReconnect(): (hostId: string) => Promise<void> {
-  const ctx = useRpcClientContext()
-  return ctx.forceReconnect
-}
-
-// Why: primes already-loaded HostProfiles so the provider can skip a second loadHosts()/Keychain pass on cold start.
-export function usePrimeHosts(): (hosts: HostProfile[]) => void {
-  const ctx = useRpcClientContext()
-  return ctx.primeHosts
 }
